fix: auth controllers

+++ DBFunctions/users/users-controllers.js
@@ -7,7 +7,7 @@ export const UsersController = (app) => {
     app.get('/users/:id', findUserById);
     app.get('/users/email/:email', findUserByEmail);
     app.delete('/users/:id', deleteUser);
-    app.put('/users/:id', updateBuyer);
+    app.put('/users/:id', updateUser);
 }
 
 const findAllUsers = async (req, res) => {
@@ -16,20 +16,34 @@ const findAllUsers = async (req, res) => {
 }
 
 const findUserById = async (req, res) => {
-    const idToFind = req.params.id;
-    const users = await usersDao.findUserById(idToFind);
-    res.json(users);
+    try {
+        const idToFind = req.params.id;
+        const users = await usersDao.findUserById(idToFind);
+        res.json(users);
+    }
+    catch (error) {
+        res.json(error);
+    }
 }
 
-const findUserByEmail = async (req, res) => {
-    const emailToFind = req.params.email;
-    const users = await usersDao.findUserByEmail(emailToFind);
-    res.json(users);
+const findUserByEmail = async (req, res, next) => {
+    try {
+        const emailToFind = req.params.email;
+        const users = await usersDao.findUserByEmail(emailToFind);
+        res.json(users[0]);
+    } catch (error) {
+        next(error);
+    }
 }
 // b. Delete -- return null when unsuccessful
 const deleteUser = async (req, res) => {
     const idToDelete = req.params.id;
-    const role =  await usersDao.findUserById(idToDelete).role;
+    const user = await usersDao.findUserById(idToDelete);
+    if (!user) {
+        res.json(null);
+        return;
+    }
+    const role = user.role;
     let status = null;
     if (role === 'buyer') {
         status = await usersDao.deleteBuyer(idToDelete);
@@ -43,22 +57,29 @@ const deleteUser = async (req, res) => {
     res.json(status);
 }
 
+
 // c. Update -- return null when unsuccessful
-const updateBuyer = async (req, res) => {
+const updateUser = async (req, res) => {
     const idToUpdate = req.params.id;
-    const updatedBuyer = req.body;
-    const role =  await usersDao.findUserById(updateBuyer).role;
+    const updatedUser = req.body;
+    const user = await usersDao.findUserById(idToUpdate);
+    if (!user) {
+        res.json(null);
+        return;
+    }
+
+    const role = user.role;
     let status = null;
     if (role === 'buyer') {
-        status = await usersDao.updateBuyer(idToUpdate, updatedBuyer)
+        status = await usersDao.updateBuyer(idToUpdate, updatedUser)
             .then(() => usersDao.findBuyerById(idToUpdate));
     }
     else if (role === 'merchant') {
-        status = await usersDao.updateMerchant(idToUpdate, updatedBuyer)
+        status = await usersDao.updateMerchant(idToUpdate, updatedUser)
             .then(() => usersDao.findMerchantById(idToUpdate));
     }
     else if (role === 'admin') {
-        status = await usersDao.updateAdmin(idToUpdate, updatedBuyer)
+        status = await usersDao.updateAdmin(idToUpdate, updatedUser)
             .then(() => usersDao.findAdminById(idToUpdate));
     }
     res.json(status);
